+++ lib/src/extension/material_state_property.x.dart
@@ -3,43 +3,43 @@
 import 'package:flutter/material.dart';
 
 extension MaterialStateProperty_TextStyleX on TextStyle {
-  MaterialStateProperty<TextStyle> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<TextStyle> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }
 
 extension MaterialStateProperty_ColorX on Color {
-  MaterialStateProperty<Color> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<Color> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }
 
 extension MaterialStateProperty_OutlinedBorderX on OutlinedBorder {
-  MaterialStateProperty<OutlinedBorder> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<OutlinedBorder> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }
 
 extension MaterialStateProperty_doubleX on double {
-  MaterialStateProperty<double> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<double> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }
 
 extension MaterialStateProperty_intX on int {
-  MaterialStateProperty<int> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<int> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }
 
 extension MaterialStateProperty_EdgeInsetsX on EdgeInsets {
-  MaterialStateProperty<EdgeInsets> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<EdgeInsets> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }
 
 extension MaterialStateProperty_BorderSideX on BorderSide {
-  MaterialStateProperty<BorderSide> get asMaterial {
-    return MaterialStateProperty.all(this);
+  WidgetStateProperty<BorderSide> get asMaterial {
+    return WidgetStateProperty.all(this);
   }
 }